Rewrite python runner PoC and use it in instance-adapter
Old version assumed the communication with the platform will happen over
unix pipes, which is no longer the case (we use TCP connections now).

The mechanism used in process-instance-adapter for choosing which runner
to use is very primitive - just checks file extension of sequence's
entrypoint.

Note that message codes and communication channel numbering is just
copied over from TS files and hardcoded.

Also, this new runner initially prints messages to a temporary log file,
so that they can be accessed before communication with host is
established.

For now, the assumption is that input and output channels will carry
text, so they are automatically decoded/encoded.

Also, absolutely no Docker support for now.