Split dropping connections and server stopping as seprate methods
Since Verser is usually used in a way that it is created based on
existing server (server is passed in a constructor) stopping the server
may break other stuff relaying on it. To make it more explicit and
hopefully less confusing there is now `disconnect()` method for dropping
connections only and `stop()` one which drops the connections and stops
the server.