Py runner: more generic LoggingSetup & server host
a) parameterize server host with env variable, with localhost default
b) abstract away the temporariness of the initial log target.
LoggingSetup should just receive a target to which it will be logging,
and optionally redirect stdout/err there (in case of process runner,
when standard streams are not accessible directly).