Capture runner logs and show in case of failure
Currently if the runner process crashes before it connects to the host
(e.g. because an import fails on missing dependency), we don't see any
logs unless DEVELOPMENT is set. This makes it hard to debug some
situations (e.g. https://app.clickup.com/t/24308805/SCP-3676).

The new approach is to save all output to a file and show it in case of
failure.