Add --exec option to run-script for non-yarn commands (#558)
...or built-in yarn commands that are not explicitly defined in
package.json, e.g. `yarn version`. For some reason @npmcli/run-script
doesn't support them (e.g. it does nothing when asked to run them).

If the command to execute has options, they should be quoted together
with the command so that they won't be interpreted as options for
run-script itself.

Also, remove --long-help as there is no long help for this script, and
don't error with usage if the user just wants the list of dirs (-l
option).

Analogous to 945235525a6f0865188719dcf2c8994e76000a5f in CPM repo.